SignalTiming - add tests for the extra railcom cutout bit
When the number of pulses for the RailCom cutout is an odd number, we need to use an extra full one bit to cover it; add a test for this

+++ Sources/DCC/SignalTiming.swift
@@ -165,7 +165,7 @@ public struct SignalTiming {
             throw Error.conformanceError(message: "Duration of RailCom cutout would be \(railComLength * pulseWidth)µs which is outside the valid range \(SignalTiming.railComRange)µs")
         }
 
-        railComCount = Int((Float(railComLength) / Float(oneBitLength * 2)).rounded(.awayFromZero))
+        railComCount = Int((railComLength / (oneBitLength * 2)).rounded(.awayFromZero))
 
         preambleCount = Self.preambleCountMin + railComCount
     }
